requirements.txt
streamlit
streamlit-authenticator>=0.2.3
pandas
PyYAML
google-generativeai
Jinja2
pytest
sqlparse
plotly>=5.0.0
altair>=5.0.0
databricks-sql-connector
reportlab
matplotlib
seaborn